Move the Delaunay filter from plugins to filters, add minimal testing
Closes #1969
