+++ plugins/CMakeLists.txt
@@ -4,10 +4,6 @@ if(BUILD_PLUGIN_CPD)
     add_subdirectory(cpd)
 endif(BUILD_PLUGIN_CPD)
 
-if(BUILD_PLUGIN_DELAUNAY)
-    add_subdirectory(delaunay)
-endif(BUILD_PLUGIN_DELAUNAY)
-
 if(BUILD_PLUGIN_GEOWAVE)
     add_subdirectory(geowave)
 endif(BUILD_PLUGIN_GEOWAVE)
