--- plugins/delaunay/CMakeLists.txt
@@ -1,11 +1,0 @@
-#
-# Delaunay plugin CMake configuration
-#
-
-#include(${PDAL_CMAKE_DIR}/geogram.cmake)
-
-PDAL_ADD_PLUGIN(filter_libname filter delaunay
-    FILES
-        filters/DelaunayFilter.cpp
-        filters/Delaunay_psm.cpp
-)
